Use the built-in interpreter for type checking tests
Share an instance of the interpreter between different tests to
save on startup costs. Import the code that's already defined in
QueryTest.scala to save on startup costs and to not duplicate.
Throw an exception if code is incomplete (this caught two bugs in
the tests already).

Test runtime went from 33s to 8s on my machine. Benefits compound
for cross-compilation.