Explicit modeling of a "command" (query + operation).
This is a significant refactoring, the initial goal of which was to enable
structured logging. We want to pass a structured command in to the log() hook
instead of a string, so we can extract useful information to log.

However it turns out that there was no explicit modeling of a command. Operations
were sometimes implicit and sometimes passed around as out-of-band strings.

This change creates a Command type hierarchy that models the various types of command.
A nice side effect of this is that the command now encapsulates much of the execution
dispatch logic, which was previous scattered all over the place in the query objects.

This is a major change, although it should not affect clients. But it provides
a cleaner more explicit model of the control flow, and should turn out to be
useful for other things besides structured logging.