compile-time index checks!
Basic idea is you do:

  Venue.useIndexN(Venue.indexName, _.indexField1, _.indexField2, ...)
       .where(_ eqs SomeValueOfField1)
       .iscan(_ in List(SomeValuesForField2))
       ...
       .fetch()

We don't technically need the indexField1,..indexFieldN parameters to
useIndexN, but I think it's necessary to keep the code at all clear,
since each individual clause doesn't name the field itself in its
predicate.

It works by having an IndexEnforcerN that exports 'where' and 'iscan'
methods that go from the next Field in the index to a QueryClause. I
also updated QueryClause's to add an IndexableQueryClause class that has
a type parameter indicating whether we think the query is indexed. I
split out a bunch of one-line classes that inherit from
IndexableQueryClause and pass the parameter appropriately.

Very rough around the edges still. Haven't figured out how this fits in
with the runtime checks. We want both for a while at least because the
runtime checks can be smarter than the compile-time checks and do checks
that are specific to Asc vs Desc vs 2d, etc., but we also don't want two
different sets of Index flags floating around that disagree with each
other.

Only works for indexes with up to 3 fields at the moment, would be easy
to make it work for more than that.