tweak syntax, add noop(), add compilation tests
Changed the syntax from:

  Venue.useIndex(Venue.idIdx, _._id)
       .where(_ eqs new ObjectId())

to:

  Venue.useIndex(Venue.idIdx)
       .where(_._id)(_ eqs new ObjectId())

This keeps the field name next to the query clause.

Also, add a noop() method so users can skip fields in an index. Added a
new phantom type to be sure that people can't noop the first field in an
index.

Added unit tests for things that shouldn't compile and tweaked the
compilation tests so you can specify a regex for the compilation error.