move actualIndexBehavior to individual QueryClause's
Rather than having a match in QueryClause that matches on the type of
the CondOp, have each QueryClause subclass pass up what their
actualIndexBehavior is.

EqClause is a little weird since it can be an index lookup, an index
scan, a partial index scan, or document scan depending on whether the
parameter is a value or a regex, and, if it's a regex, what kind of
regex it is. As a result, that needs a little bit of extra complexity.