Allow select on a subfield of an embedded list.
This is accomplished with a subselect() method on
BsonRecordListQueryField.

I tried having one subfield method on BsonRecordListQueryField that
could work in both query and select contexts, but that required that the
method return a DummyField[List[V]], which in turn meant that all
queries against a subfield of an embedded list had to use list
predicates. That was broken conceptually since it didn't match up with
the actual mongo queries, so you could've both formulated incorrect
queries and not formulated correct ones. A simple example is that if you
had a list field of embedded objects, each with an int, before you
would've been able to do:

    (_.listField.subfield(_.intField) > 5)

... but with one combined subfield method, that wouldn't have been
possible anymore.

Instead, I made a subselect method for selection only. It's a little
weird, but it's not possible to pass the output of
BsonRecordListQueryField.subfield to select() anyway, so hopefully this
won't cause many problems.

Unfortunately, this means there's this inconsistency between
BsonRecordQueryField and BsonRecordListQueryField where you use subfield
for selection on one and subselect for selection on the other.