fix for subselecting when the top-level field doesn't exist
On a select(_.foo.subselect(_.bar)), if foo was unset, rogue would throw
because it wanted foo to either be a DBObject or a BasicDBList, and it
was actually null. The fix is to just return the null and pass that off
to setFromAny. Lift's fields' setFromAny methods will correctly
deserialize that.