don't set batchSize on the cursor
Setting batchSize and sort on a cursor can cause
fewer than all of the documents to be retrieved.
Really, there's no need for the caller-facing batchSize
and the cursor's batchSize to line up. We should just
let the cursor decide what the optimal batchSize is
independent of what the caller wants to do (which
it is doing for memory reasons).