Add `setTo(vsOpt: Option[Traversable[V]])` to ListModifyField
This enables modify clauses for the form
`.modify(_.fooList setTo someOptionalList)`, where `Some(list)` will
set the field to `list` and `None` will unset the field. Rogue already
supports on this pattern on `AbstractModifyField`.

Note that this change required annotating integer literals in
`QueryTest.scala` as being of type `Long`, as it seems to break the
implicit conversion for widening `Int` to `Long` in the query. I have no
opinion on whether the enhanced functionality in this patch justifies
the break in backwards compatibility.